Decouple Node implementations and IO port collections
Originally an implementation has to call
    this.inputCollection.add(this.input)
and
    this.outputCollection.add(this.output)
to add them to the input and output port collection of the node and let them be executed by the external Statement.
However, this is completely not necessary, as reference of the node itself is passed in during the construction ( instantiation ) of the input / output ports, so that the node can be accessed using the ports' getParent() method.
This is for the Statement.execute() method and a number of other ones , so that two adjacent nodes ( connected by an Edge ) can refer each other through their IO ports, therefore the whole node tree can be traversed starting from one single input port.
Then came the thoughts that why can't IO ports just claim their existence to the node just by themselves? They have their parent nodes' reference, shouldn't that be easier and more logical?
While not doing this may confuse people new to this node-graph model, them may omit these calls and wonder why their code don't work.

So now is it implemented and the problems are solved.
And, test passed.