Remove PointsWhiteNoise as ISupplier
In my project all random components should be pseudo-random, which returns the same value ( actually, because of the wide usage of parallel streams, the result merely *contains the same thing*, the order of which is not guaranteed ) with the same seed. Therefore, PointsWhiteNoise should also take a seed as input, instead of simply using currentTimeMillis() as seed. Besides, enabling PointsWhiteNoise to be casted both to ISupplier and IProcessor would result in ambiguous call with NodeGraphHelper.create().